requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-clang-format==0.12.11
ament-cmake-test==1.3.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
bond==4.1.2
builtin-interfaces==1.2.1
cachetools==6.2.2
can-msgs==2.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
composition-interfaces==1.2.1
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-aggregator==4.0.2
diagnostic-msgs==4.2.4
diagnostic-updater==4.0.2
distro==1.9.0
docstring_parser==0.17.0
domain-coordinator==0.10.0
emojis==0.7.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.1
filelock==3.20.0
fsspec==2025.12.0
generate-parameter-library-py==0.4.0
geodesy==1.0.6
geographic-msgs==1.0.6
geometry-msgs==4.2.4
google-auth==2.43.0
google-genai==1.54.0
gps-msgs==2.0.4
grid-map-msgs==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
interactive-markers==2.3.2
Jinja2==3.1.6
jiter==0.12.0
laser-geometry==2.4.0
launch==1.0.7
launch-ros==0.19.8
launch-testing==1.0.7
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
lazy_loader==0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.3
message-filters==4.3.5
mpmath==1.3.0
nav-msgs==4.2.4
nav2-common==1.1.18
nav2-msgs==1.1.18
networkx==3.4.2
nmea-msgs==2.0.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.9.0
osrf-pycommon==2.1.4
packaging==25.0
pacmod3-msgs==1.0.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-qt-binding==1.1.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
radar-msgs==0.2.1
rcl-interfaces==1.2.1
rclpy==3.3.15
rcutils==5.1.6
requests==2.32.5
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.13
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.8
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2test==0.4.0
ros2topic==0.18.11
rosapi==2.0.0
rosapi-msgs==2.0.0
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosbridge-library==2.0.0
rosbridge-msgs==2.0.0
rosbridge-server==2.0.0
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-monitor==1.0.6
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
septentrio-gnss-driver==1.4.1
shape-msgs==4.2.4
smclib==4.1.2
sniffio==1.3.1
sros2==0.10.5
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
theora-image-transport==2.5.3
tifffile==2025.5.10
topic-monitor==0.20.5
topic-tools==1.1.1
topic-tools-interfaces==1.1.1
torch==2.9.1
tqdm==4.67.1
trajectory-msgs==4.2.4
triton==3.5.1
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ublox-msgs==2.3.0
udp-msgs==0.0.5
unique-identifier-msgs==2.2.1
urllib3==2.6.1
velodyne-msgs==2.5.1
visualization-msgs==4.2.4
websockets==15.0.1
xacro==2.0.8
